backend/requirements.txt
fastapi>=0.121.2
nats-py>=2.12.0
regex>=2025.11.3
uvicorn>=0.38
python-multipart==0.0.20
polars>=1.35.2
aiofiles==25.1.0